feat: created component folder for uber
